@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
@@ -9,14 +10,14 @@ namespace BareBonesCRUDMicroservice.Model
     public class BareBonesCRUDItem : BaseEntity
     {
         //TODO: Change this to a type EF accepts. For now I'm ignoring it.
-        [NotMapped]
-        public List<string> Items { get; set; }
+        public List<SubBareBonesCRUDItem> Items { get; set; }
+        [Required]
         public string Name { get; set; }
         public string Status { get; set; }
 
         public BareBonesCRUDItem()
         {
-            Items = new List<string>();
+            Items = new List<SubBareBonesCRUDItem>();
         }
     }
 }
